requirements.txt
ipaddress
google-cloud-dns==0.35.0
google-auth==2.38.0
whatismyip
loki-logger-handler
